proxy: mux cleanup and bug fix. (#399)
Background:
http handlers that require WebSockets need to be passed an http.RequestWriter
that can be type casted into an http.Hijacker.
This dependency is not well documented.

The proxy code further requires a Mux that can match host names.
Given that the default http.Mux cannot match on hosts out of the box,
back when the code was written I ended up using muxie (popular and
very performant, used by the iris framework).

When testing nassh, I realized that the gorilla library for WebSockets
used by nassh was not working, but I could not pinpoint the exact issue.

To work around the problem I ended up using a hack where I combined multiple
muxes together: a simple domain mux (in khttp, hand written), a default
http.Mux, and muxie. This allowed nassh to work.

This, however, had a couple of undesireable side effects:

1) Code complexity: it is hard to reason about the code, the
   mux nesting is just bad.
2) Breaking WebSockets for normal proxied connections.
   Turns out that the default golang proxy implementation also has a dependency
   on being able to use http.Hijacker on the http.RequestWriter
   in order to provide (supported) WebSockets.

In this PR:
1) Document the requirements of the mux used by enproxy through
   a new interface file, amux.
2) Added a generic adapter that gets muxie to work correctly in
   all case, thanks to a simple typecast when invoking the handler.
3) Greatly simplified the enproxy code by using a single mux,
   everywhere.

Tested:
a) A single mux is now used everywhere. If websockets are broken by
   the mux, the nassh tests fail. All tests are passing, and verified
   that with the typecast removed, the test hangs and eventually
   fails (retry strategy in trying to create a WebSocket, until it
   fails).
b) A variation of this PR has been running on our proxy for about
   a week for qualification purposes. I can see websockets via nassh
   and grafana working correctly.