third_party: Import BuildBuddy protos (#411)
This change uses Copybara to selectively import a few BuildBuddy protos
needed to make HTTP RPC calls to fetch BES logs by invocation ID.

Copybara is used to easily modify the protos as they are imported. This
is necessary because even though BuildBuddy is itself a bazel workspace,
it includes copied protos from the bazel repository that conflict with
the same protos already imported under //third_party/bazel. Using
Copybara allows for the import paths to be rewritten as they are
imported.

This change squashes the first import into the same change that adds the
Copybara script, BUILD files, and README.

Jira: INFRA-468