proxy/nasshp: correctly handle close request.
Background:
The API allows users of ptunnel to, well, close the tunnel when desired.
In the current implementaion if a tunnel is closed without error,
the library will keep trying to connect anyway.

In this PR:
- make sure errors are propagated correctly, implement Unwrap in one
  of the internal error types.
- have a well defined error returned by close, so the rest of the
  code can do the right thing.

Tested:
end to end test (in another PR) does not timeout, Close() results in clean
shut down.