Bazel FUSE: Add kzip library
This change adds a library to assist in unpacking ZIP archives to
temporary directories. This will be useful in the Bazel FUSE
implementation to open and reroot `output.zip` archives from bazel tests
easily.

This implementation requires that the ZIP file be effectively downloaded
in its entirety, as it spills all the contents to disk. While this won't
scale for large files due to network bandwidth, latency, and disk space
concerns, it should be useful enough for small outputs while we
determine the scalable solution for ZIP archives of all sizes.

This change also moves test helper functions out to the testutil
library:
* `assertCmp` from flextape service tests becomes `testutil.AssertCmp`
  (though this was later dropped from `kzip` tests)
* `testutil.MustRunfile` is added, which finds a file under the runfiles
  directory or panics.

Jira: INFRA-472