proxy/nasshp: make the code race free. (#420)
* proxy/nasshp: make the code race free.

Background:
go support a great race detector. By turning on the race detector
for all the tests in proxy/, it detected a couple issues in browser.go.

This PR:
- addresses all detected races in the proxy/ codebase, by changing
  rack/wack variables to use atomics, which was the initial intent
  when the code was written.

Tested:
Before this change, running the tests for enproxy with the race
detector on reported problems in browser.go. After this change,
no races are detected.

* Addressed Scott's comment.