presubmit: Move to canonical dev image
Now that the developer image has a recent version of enkit, we can
migrate off of the custom version with the patched version of enkit
(originally done to get a version of enkit with the bazel presubmit
subcommands).

This change causes the presubmit to use whichever image is tagged
`enkit-bazel-presubmit`, which will allow us to update the presubmit
container without code changes in the future (build reproducibility
concerns are relaxed around the presubmit itself, so no need to pin to
specific image hashes).

Jira: INFRA-111