bestie: Add test metrics upload to BigQuery (#423) (#429)
* bestie: Add test metrics upload to BigQuery (#423)

This continues the work of #394 and #406 by adding support
for uploading test metrics to BigQuery. The changes in this
PR do the following:

- Read and process metrics contained in *.metrics.pb file(s)
produced during bazel test runs. Format and push these
test metrics to the BigQuery 'bestie-builds.<dataset>.testmetrics'
table, where <dataset> is specified by the --dataset command line
parameter.

- Supports both file:// and bytestream:// schemes for the
outputs.zip file in the TestResult event, which correspond to
local and cluster operation of the BES Endpoint, respectively,
using APIs provided by enkit/lib/kbuildbarn. A new --baseurl
command line parameter allows a deployment-specific base URL
to be specified.

- Adds several service metrics related to BigQuery interactions
for monitoring the BES Endpoint operation.

Tested: Ran BES Endpoint service both locally and in the staging
cluster, where it detected the test.metrics.pb file produced by
a standalone test app. Authentication issues prevented the staging
cluster service from communicating with BigQuery, but that will be
resolved in a separate PR by creating a new service account.

Jira: INFRA-423

* bestie: Addressed code review comments

Many changes in this commit to respond to the review comments
in this PR. The most noteworthy of these are:
- No longer creating non-existent datasets or tables.
- Replaced all console print messages with a logger.
- Added --debug support for additional log messages.
- Revamped service stats implementation using counter IDs.
- Using a goroutine to update Prometheus metrics.