Bazel FUSE: Propagate correct port to bb_clientd
Prior to this change, bb_clientd was started with a hardcoded port,
instead of the port of the actual service or the tunnel, if one
was opened.

This change removes the hardcoded port and passes the correct port to
bb_clientd initialization.

Tested:

```
bazel run //enkit:enkit -- \
  outputs \
  mount \
  --invocation=6175ddd6-de64-447d-a827-87bba04391f4 \
  --url=<REDACTED> \
  --api-key=<REDACTED> \
  --buildbarn_host=<REDACTED>

cat ~/outputs/6175ddd6-de64-447d-a827-87bba04391f4/test.log
```

Jira: INFRA-518