bestie: Use zipstream package to read outputs.zip
Change the handling of the outputs.zip file so that it is not
read into memory prior to processing each of its file elements.
Use the https://github.com/xenking/zipstream package to iterate
through each of the files, looking for file names ending in
.metrics.pb.

Tested: Ran BES Endpoint service both locally and in the staging
cluster, where it detected the test.metrics.pb file produced by
a standalone test app.

Jira: INFRA-504