Bazel FUSE: Separate request host from tunnel endpoint
When using a tunnel, two hostnames are requried:
* The host to tunnel to (`pbzl.service.enfabrica.net`)
* The host to make requests to (`build.local.enfabrica.net`)

The latter is required because the tunnel host is typically a
non-routable internal IP, whereas this host resolves to localhost if a
tunnel is required; additionally, ingress will route requests based on
the destination host, which is expected to be the build.local... host.

This change breaks the single host flag into two flags: One is used to
resolve to a non-localhost address (if no tunnel is required) or a
localhost address (if a tunnel is required), and requests are to be
directed to this URL. The new flag specifies the endpoint to tunnel to
if a tunnel is required.

Tested:

Jira: INFRA-518