enkit outputs: Don't sleep waiting for bb_clientd (#453)
* enkit outputs: Don't sleep waiting for bb_clientd

This change replaces the 5-second sleep waiting for bb_clientd to start
and create the mount with a polling loop that looks for an external
signal that bb_clientd is ready. The external signal chosen is that one
of the expected mount directories exists. The code will poll for up to 5
seconds instead, which should allow it to move on more quickly once
bb_clientd is ready

Tested: Ran:

```
bazel run //enkit -- shutdown
bazel run //enkit -- outputs mount \
        --buildbarn_host=<redacted> \
        --buildbarn_tunnel_target=<redacted> \
        --api_key=<redacted> \
        --buildbuddy_url=<redacted> \
        --invocation=6c01fa88-3fe5-47e8-8a5a-6fb17b2330eb
```

which mounted successfully much more quickly.

Jira: INFRA-547

* Move to retry library

* Shutdown process in error paths

* Convert pollForProcessEnd to retry lib

* Tweak error handling

* fix build issue