flextape: introduce prioritization support.
Background:
Before this PR, licenses were always allocated in FIFO order.

In this PR:
- allow to specify a prioritization strategy (per license).
- define an interface to provide different prioritization strategies,
  refactor the code to allow for the plumbing.
- create a strategy that spreads licenses evenly across users.
- added extensive tests.

Note that this PR still does not allow to allocate multiple licenses
at once. The rough plan to allow for that is to expand the concept
of "invocation", so it links to all the types it is trying to allocate,
and extend the prioritizer to account for that.

For now, I just wanted to allow better license spreading, and figure out a
reasonable interface for queue reordering.

Tested: added significant unit tests, all passing.