marshal: Fix type detection when passed a URL
When a YAML config is included in another config today by URL, the
config language detection by extension fails because it is passed a URL
that has a query string suffix instead of a local file path. When this
happens, the default unmarshaler of JSON is used, which causes obtuse
parse errors to be thrown when parsing other formats.

This change first tries to parse the input path as a URL, falling back
to the assumption that the path is a local file path if URL parsing
fails. URL parsing allows us to pick out the path component, ignoring
all other components like query strings that may suffix the string.

Tested: Added unit test to demonstrate issue and verify fix.

Jira: INFRA-487