flextape: Add license release type metric
This change adds a metric that tracks the count of license releases
across different reasons, including when licenses expire. This is
intended to help debug an issue where errors are being returned by
Flextape for unknown invocations.

If we see metrics indicating that these invocations are expiring prior
to these errors being returned, we may be able to narrow down the
problem. In general, it will be useful to see how often expirations
occur, as ideally this would be close to 0.

Jira: INFRA-607