bestie: Remove support for unstructured text.xml files
Decided not to support test.xml files that are unstructured
XML, namely the individual test case results are buried within
console text output that must be scraped to create pass/fail
metrics.

Changed error handling related to XML file parsing and structure
such that it does not propagate an error back up the call chain.
Instead, an appropriate counter is incremented and a debug log
message is written.