faketree: wait for all child processes to die before returning. (#482)
Before this PR:
faketree would wait for its direct child to die, and then exit with its exit
status. If this child process spawned other children, it would not wait for
them to complete.

This is a problem when the parent, for example, is sending SIGTERM to the
entire process group - terminating the direct children possibly sooner than
indirect children, and returning/completing status before all are killed.

In this PR:
Add code so faketree _waits_ for its own child, and any indirect child
spawned within the namespace.

This is protected around the --wait= option, turned on by default, but
can easily be disabled with --wait=false.

Tested: simple test added in faketree_test.sh. Manually tested with
commands like:
	faketree -- sh -c "(sleep 5 & (sleep 2 & sleep 5) & sleep 7) &"

To verify that all commands are waited for.

This is another step toward a workaround for INFRA-256 and INFRA-535.