bestie: Add 'name' and 'created' metrics tags support
To facilitate Grafana dashboard table displays, copy the
metricname into a new 'name' tag so that it appears among
the list of displayable entities for Grafana.

Support a new 'created' tag that contains a timestamp to
be used for the metric instead of the one defined as the
original timestamp field. This allows the test application
to define a single value timestamp that is stored in the
'created' tag for both the pass/fail testresult metric and
any associated data metrics generated by the test cases
during the same run. Grafana can then use the 'created'
tag (instead of the Prometheus scrape timestamp) as a
consistent way to filter all metrics for a given time
selection.

Tested: Ran `bazel test --cache_test_results=no --test_output=streamed \
--bes_backend=grpc://localhost:6433 //systest/src/demo:test_sample_metrics'
to produce some sample metrics and verified the existence of
the new tags. Was able to set up Grafana queries using the
'created' tag to display both the test result and related
data metrics using various timestamp selections.

Jira: INFRA-744