affected targets: Pass if build graph is fixed between start and end (#495)
Presubmit failing with PRs that fix a build graph breakage are
confusing and cause delays in landing the fix PRs. At the same time,
there is no sensible set of affected targets to build if the build graph
cannot be queried at both points, which it can't if it was broken at one
of them.

This PR adds a special-case for runs where the build graph was broken at
the start commit (as determined by the entire query failing for any
reason) and fixed at the end commit - the library returns no changed
targets and no error. This should cause the presubmit to pass with a
no-op build in this case.

This is still not a very good presubmit because it doesn't build
anything, and a warning is emitted to this effect. However, it will at
least reduce confusion with fix PRs.

Tested: not at all

Jira: INFRA-174