khttp: fix server.go to be more reasonable.
Background:
server.go was created out of the proxy in a hurry, turns out it's more
broken than I'd like to admit. In this specific case:
- flags were not prefixed properly
- description were not updated
- semantics were not intuitive

Last point bit me last night: I was assuming the port number to be
a default if not specified in the address, turns out that if an address
is specified at all, the port number is completely ignored.

In this PR:
- fix the prefix
- fix the description
- have the port number behave as a true default port number

Additionally, added small unit test, and go fmt.