BES Endpoint: Expand test.xml search for "realm" tag
To support multiple test client environments, such as pytest and
gtest, the location of a "realm" tag may not always appear in the
test case properties. This change looks in the attributes of the
<testsuite> and <testsuites> elements of the test.xml file, in
that order, as additional sources of the "realm" tag for the
'testresult' metrics.

Tested: Used variations of `bazel test //model/functional/sfa-model:model-main-test`
to place the "realm" tag in different locations within the test.xml
file. Verified by looking at the console log output of a local bestie
instance that bestie found the tags and correctly prioritized
the one to use in the 'testresult' metric sent to BigQuery.

Jira: INFRA-826