astore: Add unit tests for server.Retrieve
This change aims to add unit tests for Server.Retrieve, so that behavior
can be safely modified in a future change. It does this by replacing the
`ds` field of `Server` with an interface that has the required subset of
methods from `datastore.Client`, and then providing a fake in the tests
that will record the queries made for later comparison and assertion.

Jira: INFRA-830