astore client: Specify empty tag set when querying by ID
Currently, when querying by UID, the client sends a request with no tags
to the server, which the server interprets by defaulting a latest tag
before running the query. This causes the query to return no results
when the package with that UID is not marked `latest`.

With this change, the client will always send a request with an empty
`TagSet`, which the server interprets as "set no default". This allows
queries for specific UIDs to proceed as expected.

Tested: TODO

Jira: INFRA-830