astore client: Specify empty tag set when querying by ID (#522)
Currently, when querying by UID, the client sends a request with no tags
to the server, which the server interprets by defaulting a latest tag
before running the query. This causes the query to return no results
when the package with that UID is not marked `latest`.

With this change, the client will always send a request with an empty
`TagSet` when querying by UID, which the server interprets as "set no 
default". This allows queries for specific UIDs to proceed as expected.

Tested: `bazel run //enkit -- astore download --force-uid rqh8i6qspag3ynxdqdu4k33keymgmxkn --output=/tmp/test_ip.tgz --loglevel-console=debug --overwrite` now downloads the requested file instead of erroring

Jira: INFRA-830