enproxy: add automated session expiration.
Background:
When a client disconnects, enproxy does not close the session between
the proxy and the backend, so if the client comes back, the session
can be resumed.

This causes connections to build up over time, and pretty much leak
memory until the connections naturally die.

In this PR:
- add a few counters to keep track of orphaned connections, export
  them to prometheus for monitoring.
- use those counters to periodically and automatically expire sessions.

Additionally, a few support classes were needed (and added):
- atomic time, to keep track of time atomically.
- a mock clock implementation, to abstract time.After.

It's a large PR, it could probably have been broken in two,
but it's hard to do after the fact.

Tested: written unit tests, will need to stage the change.