kunit: Fix BUILD files, imports
This change fixes Python import paths and BUILD files for the kunit
binary/library to make it Bazel compliant.

An extra target is also defined to build a Python hermetic zip; this zip
file is ran instead of the script, so that we don't have to worry as
much about runfiles issues. The script template is modified to pass the
zip to python3, as the zip is not itself executable.

Tested: `BAZEL_PROFILE=buildbarn bazel test
//driver/farmem:enf_rma_unit_test
--override_repository=enkit=/home/bminor/dev/enkit` with internal PR
1656 works