enkit bazel: Log error when start query broken
When the start query is broken but the end query is not, we may be in a
situation where the build graph is broken on master, and this presubmit
is for a fix PR, so we want to ignore the build graph breakage at the
start point.

However, for PRs that aren't fixing build graph breakage, we are
potentially passing the presubmit without actually testing anything.

This change logs the error that was encountered when evaluating the
build graph, to allow users to see and debug the underlying issue
causing query to fail.

Jira: INFRA-899