flextape: Raise and parameterize timeouts
All durations that the service uses are moved to a "server" config
section, so that they are no longer hard-coded. At the same time,
timeouts are greatly raised to reduce the number of expirations that
happen because the client is too slow to refresh an active allocation.

The client is modified to attempt a refresh earlier into the
server-specified duration, but given the raised timeouts its minimum
refresh duration is raised from 1s to 5s.

Jira: INFRA-607