kbuild/utils/buildroot: add init.sh support to run_kunit_tests.sh. (#553)
Background:
I'm changing how uml tests are run. Specifically, in the next PR
bazel will generate a "runtime package" - a directory with data and
an init.sh script to "run the test" (or really, anything) in a
generic VM (or UML).

In this PR:
- update run_kunit_tests.sh so that it checks for the presence
  of an init.sh script in the mounted directory, and if it is there,
  runs it.

- while at it, various improvements, specifically:
  - make the script `shellcheck` clean. This required a lot of
    style improvements: error handling, " in places we did not
    have before, etc.

  - exit calls required invoking poweroff reliably, so moved
    to a trap event, and moved to `poweroff -f`, as to avoid
    going through systemd/init (faster, more reliable)

  - removed some obsolete / incorrect comments, added some
    new ones.

- Additionally, added instructions on how to edit an existing
  image without having root or `mount -o loop` available.

Tested:
I actually released a new version of the image with the improved
script as the only change in the rootfs. All tests in internal
are passing.