bazel/linux: split defs.bzl in multiple files.
Background:
defs.bzl has grown too large. Upcoming changes are adding support
for kvm, which will make it even larger.

In this PR:
- create a providers.bzl - must to avoid circular dependencies.
- moved uml specific code into dedicated file - uml.bzl.

This change is a noop: no code has been harmed in the process.

Tested:
bazel test with --override_repository in internal shows no
regression in the tests.