enkit ssh: Add --proxy-map arg
This change adds a `--proxy-map` argument to `enkit ssh`, allowing one
to map bits of a SSH target address to its corresponding proxy. This
will allow us to configure e.g.:

```
--proxy-map=.foo=https://foo.example.com,.bar=https://bar.example.com
```

in the enkit kconfig, so that a bare `enkit ssh user@machine-1.bar` will
resolve to the (potentially non-default) proxy https://bar.example.com.

This is done by matching all of the non-flag SSH args against all
entries in `--proxy-map` via substring lookup until a match is found; if
no match is found, the proxy named by `--proxy` is used as a fallback.

This logic is likely to work most of the time but fall on its face for
unusual invocations of `enkit ssh`, so proxy decisions are logged at at
least the INFO level.

Jira: INFRA-868