WORKSPACE: Completely reorganize
This change reorganizes WORKSPACE into multiple Starlark files, which
both enkit and downstream repositories can call, ensuring that
dependencies between the two are identical.

This change serves as the basis for moving dependencies from internal
into enkit as appropriate, and updating dependencies as needed to ensure
that both enkit and internal are up-to-date.

WORKSPACE files are complicated, and this change is organized into
multiple Starlark files instead of one for technical reasons that are
captured in an associated README.

Jira: INFRA-1038