kbuild/v2: Update scripts to build/publish UML kernel artifacts
This patch updates the kbuild scripts to also build and publish the
UML kernel artifacts.

The main runner script is now called "kpub-astore" and has been
reworked to take command line arguments.

This patch also updates the README to describe how to use the script
and what the scripts do.

Fixes: INFRA-952

Signed-off-by: Curt Brune <curt@enfabrica.net>