bazel/linux: turn {uml|qemu} test rules into runnable rules.
Background:
So far, all the uml_|qemu_ virtualization rules were for testing
only. Now we want to use them just to run an emulator.

This PR changes the rules from "test = True" to "executable = True",
with all that it implies.

Specifically:
- refactor the .sh scripts - they can no longer rely on TEST_
  environment variables being set.

- introduce an exec_test.bzl utility - that can turn any executable
  rule into a test rule.

- use said rule in existing definitions so to maintain backward
  compatibility with the existing ruleset.

- have the shell scripts in the rules be more useful. Specifically:
  - it is now possible to override emulator/kernel/rootfs parameters
    directly from the command line.

  - paths and runtime info is clearly displayed on the screen, removing
    sandboxing parameters, so it's very easy to debug those scripts
    manually.

Tested:
Existing regression tests are still passing in internal/. An example
PR in qemu is failing exactly as it was before :) needs the module
loading plumbing to be useful.