astore: Fix bug with astore_download_and_extract implementation
A repository rule's implementation only gets the ctx object as a
parameter, and this commit fixes that with a wrapper function. In a
perfect world, we would pass any attribute needed in via ctx.attr, but
like noted in the inline comment, kernel_tree_version uses "sha256" in
place of "digest", and changing it now would break external consumers of
the rule and a rolling update will not be feasible. Longer-term, this
wrapper should be replaced with a dict of attributes for the various
consumers of the common function.

The commit also introduces a strip_prefix attribute for the general
astore_package rule.

Tested manually with both astore_package consumers (which was missed
with cdf318cf0) and kernel_tree_version consumers (with and without an
authenticated download).

Signed-off-by: Raghu Raja <raghu@enfabrica.net>