bazel/linux: further improve bundles. (#600)
With this PR, a runner can take an arbitrary set of runners.
Further, runners can be simple bazel target, without need
of custom rules.

Basically, to load kernel modules, now we can have a:

    sh_binary(
        name = "prepare-all-modules",
        srcs = [
          "script.sh",
        ],
        data = [
          "@kernel-modules//:all"
        ],
    )

where all script.sh does is "mount --bind ./lib/modules /lib/modules && depmod -a",
and we can list to run in a target. Something like:

    kernel_qemu_run(
       ...
       run = [
           ":init-system",
           ":prepare-all-modules",
           ":run-kunit-test",
       ],
       ...
    )

In this PR:
- break down the KernelRuntimeInfo into two providers: one
  representing a set of actions (what to do to prepare the system,
  to check the system, to run in the vm), and specific action.
- improve runner so it can take any set or list of actions.
- minimal backward compatible changes to defs.bzl, so old
  tests keep running.

This is yet not very ergonomic, a couple fixes are coming,
including an update in internal.