enkit: remember last known good DNS config endpoint (#604)
enkit: remember last known good DNS config endpoint

Some enkit commands are dynamically loaded over the network. The config
files and other bits pertaining to these commands are cached locally so
that the commands can still be used if the network is later unavailable.

The location of the configuration file for the domain, which is also used for
the subsequent cache lookups, is retrieved via DNS. The result of the
DNS lookup, however, was not cached. The net effect of this is that
even though the commands and their config were indeed cached, there was
no mechanism to find them in the absence of DNS.

Storing a previously good fallback configuration endpoint and using this
when the DNS query fails allows the cached commands to work as expected
when disconnected from the network.

Bug: INFRA-970
Signed-off-by: Edwin Peer <edwin@enfabrica.net>