INFRA-1088: bestie: enforce 20 MB output file size limit
The BES Endpoint will not attempt to read and process any output
file whose size exceeds 20 MB. This is to prevent out-of-memory
conditions being experienced several times per day in the production
cluster. This does not impact the running 'bazel test' job, but
will preclude updating the BigQuery database with any metrics
contained in the unprocessed files.

Note that the built in default of 20 MB for the file size limit
can be overridden using the --max_file_size command line
parameter.

Tested: Ran a local instance of the bestie server along with
a bazel test //systest/src/demo:test_sample_metrics that sent
its BES events to localhost:6433. Observed the desired behavior
when files exceeded the limit.

Jira: INFRA-1088