INFRA-1088: Use LimitReader to restrict ReadAll
Use a LimitReader to impose a limit on the number of bytes that
ReadAll reads. Output files larger than this limit are not
processed by bestie.

Tested: Determined that a particular test.xml file was 554 bytes.
Set the command line --max_file_size parameter to 554 and saw
that the file was processed. Tried again at 553 and saw the debug
message that the file was too large.