INFRA-1088: bestie: enforce 5 MB output file size limit (#623)
* INFRA-1088: bestie: enforce 5 MB output file size limit

The BES Endpoint will not attempt to read and process any output
file whose size exceeds 5 MB. This is to prevent out-of-memory
conditions being experienced several times per day in the production
cluster. This does not impact the running 'bazel test' job, but
will preclude updating the BigQuery database with any metrics
contained in the unprocessed files. Note that the built in default
of 5 MB for the file size limit can be overridden using the
--max_file_size command line parameter.

Instead of checking the file size using os.stat(), which likely will
not work for a bytestream file, use a LimitReader to impose a
limit on the number of bytes that ReadAll reads. Output files
larger than this limit are not processed by bestie.

Tested: Ran a local instance of the bestie server along with
a bazel test //systest/src/demo:test_sample_metrics that sent
its BES events to localhost:6433. Observed the desired behavior
when files exceeded the limit.

Manually determined that a local test.xml file was 554 bytes.
Set the command line --max_file_size parameter to 554 and saw
that the file was processed. Tried again at 553 and saw the debug
log message that the file was too large.

Jira: INFRA-1088